Move CountryView to views v2 in users app
