+++ EquiTrack/users/views_v2.py
@@ -1,4 +1,12 @@
+import logging
+
+from rest_framework.generics import ListAPIView
+
 from users import views as v1
+from users.models import Country
+from users.serializers import CountrySerializer
+
+logger = logging.getLogger(__name__)
 
 
 class UserAuthAPIView(v1.UserAuthAPIView):
@@ -13,8 +21,22 @@ class StaffUsersView(v1.StaffUsersView):
     """Stub for StaffUsersView"""
 
 
-class CountryView(v1.CountryView):
-    """Stub for CountryView"""
+class CountryView(ListAPIView):
+    """
+    Gets a list of Unicef Staff users in the current country.
+    Country is determined by the currently logged in user.
+    """
+    model = Country
+    serializer_class = CountrySerializer
+
+    def get_queryset(self):
+        user = self.request.user
+        if not user.profile.country:
+            logger.warning('{} has not an assigned country'.format(user))
+            return self.model.objects.none()
+        return self.model.objects.filter(
+            name=user.profile.country.name,
+        )
 
 
 class MyProfileAPIView(v1.MyProfileAPIView):
